// packages/tokens-studio-for-figma/src/types/values/TokenTypographyValue.ts
export type TokenTypographyValue = {
  fontFamily?: string;
  fontWeight?: string;
  fontSize?: string | number;
  lineHeight?: string | number;
  letterSpacing?: string | number;
  paragraphSpacing?: string | number;
  paragraphIndent?: string | number;
  textCase?: string;
  textDecoration?: string;
};